+++ Charts/Classes/Data/BubbleChartDataEntry.swift
@@ -10,6 +10,7 @@
 //
 
 import Foundation
+import CoreGraphics
 
 public class BubbleChartDataEntry: ChartDataEntry
 {
@@ -19,7 +20,7 @@ public class BubbleChartDataEntry: ChartDataEntry
     /// :xIndex: The index on the x-axis.
     /// :val: The value on the y-axis.
     /// :size: The size of the bubble.
-    public init(xIndex: Int, value: Float, size: CGFloat)
+    public init(xIndex: Int, value: Double, size: CGFloat)
     {
         super.init(value: value, xIndex: xIndex)
         
@@ -30,7 +31,7 @@ public class BubbleChartDataEntry: ChartDataEntry
     /// :val: The value on the y-axis.
     /// :size: The size of the bubble.
     /// :data: Spot for additional data this Entry represents.
-    public init(xIndex: Int, value: Float, size: CGFloat, data: AnyObject?)
+    public init(xIndex: Int, value: Double, size: CGFloat, data: AnyObject?)
     {
         super.init(value: value, xIndex: xIndex, data: data)
