+++ Charts/Classes/Utils/ChartFillFormatter.swift
@@ -12,12 +12,12 @@
 //
 
 import Foundation
-import CoreGraphics.CGBase
+import CoreGraphics
 
 /// Protocol for providing a custom logic to where the filling line of a DataSet should end. If setFillEnabled(...) is set to true.
 @objc
 public protocol ChartFillFormatter
 {
     /// Returns the vertical (y-axis) position where the filled-line of the DataSet should end.
-    func getFillLinePosition(#dataSet: LineChartDataSet, data: LineChartData, chartMaxY: Float, chartMinY: Float) -> CGFloat;
+    func getFillLinePosition(#dataSet: LineChartDataSet, data: LineChartData, chartMaxY: Double, chartMinY: Double) -> CGFloat;
 }
